# feat(functions): converting all aggregates to use arrow arrays