--- a/functions/tests/testdata/count.out.csv
+++ b/functions/tests/testdata/count.out.csv
@@ -1,0 +1,10 @@
+#datatype,string,long,string,string,long
+#group,false,false,true,true,false
+#default,_result,,,,
+,result,table,_measurement,_field,_value
+,,0,iZquGj,Au1iY,6
+,,1,iZquGj,HSbYC,6
+,,2,iZquGj,J1u,6
+,,3,iZquGj,ei77f8T,6
+,,4,iZquGj,ucyoZ,6
+