--- a/functions/tests/testdata/mean.flux
+++ b/functions/tests/testdata/mean.flux
@@ -1,10 +1,7 @@
-t_mean = (table=<-) =>
-  table
-    |> range(start:2018-05-22T19:53:26Z)
-    |> group(columns:["_measurement", "_start"])
-    |> mean()
-    |> map(fn: (r) => ({mean: r._value}))
-    |> duplicate(column:"_start", as: "_time")
-    |> yield(name: "0")
+option now = () => 2018-12-18T22:12:00Z
+
+t_mean = (table=<-) => table
+  |> range(start: -5m)
+  |> mean()
 
 testingTest(name: "mean", load: testLoadData, infile: "mean.in.csv", outfile: "mean.out.csv", test: t_mean)