--- a/functions/tests/testdata/percentile_tdigest.out.csv
+++ b/functions/tests/testdata/percentile_tdigest.out.csv
@@ -2,7 +2,5 @@
 #group,false,false,true,true,false
 #default,_result,,,,
 ,result,table,_measurement,_field,_value
-,,0,Sgf,DlXwgrw,48
-,,1,Sgf,GxUPYq1,9.5
-,,2,Sgf,qaOnnQc,10.979094987144862
+,,0,SOYcRk,NC7N,29.50336437998469