--- a/functions/tests/testdata/skew.out.csv
+++ b/functions/tests/testdata/skew.out.csv
@@ -2,7 +2,7 @@
 #group,false,false,true,true,false
 #default,_result,,,,
 ,result,table,_measurement,_field,_value
-,,0,Sgf,DlXwgrw,48
-,,1,Sgf,GxUPYq1,9.5
-,,2,Sgf,qaOnnQc,10.979094987144862
+,,0,Sgf,DlXwgrw,0.3057387969724791
+,,1,Sgf,GxUPYq1,0.7564084754909545
+,,2,Sgf,qaOnnQc,0.6793279546139146