--- a/functions/tests/testdata/stddev.out.csv
+++ b/functions/tests/testdata/stddev.out.csv
@@ -2,7 +2,7 @@
 #group,false,false,true,true,false
 #default,_result,,,,
 ,result,table,_measurement,_field,_value
-,,0,Sgf,DlXwgrw,48
-,,1,Sgf,GxUPYq1,9.5
-,,2,Sgf,qaOnnQc,10.979094987144862
+,,0,Sgf,DlXwgrw,20.5621010599598
+,,1,Sgf,GxUPYq1,52.060541679855774
+,,2,Sgf,qaOnnQc,59.9881468845287