--- a/functions/transformations/count_test.go
+++ b/functions/transformations/count_test.go
@@ -5,10 +5,12 @@ import (
 	"time"
 
 	"github.com/influxdata/flux"
+	"github.com/influxdata/flux/arrow"
 	"github.com/influxdata/flux/execute"
 	"github.com/influxdata/flux/execute/executetest"
 	"github.com/influxdata/flux/functions/inputs"
 	"github.com/influxdata/flux/functions/transformations"
+	"github.com/influxdata/flux/memory"
 	"github.com/influxdata/flux/querytest"
 )
 
@@ -83,10 +85,11 @@ func TestCount_Process(t *testing.T) {
 	)
 }
 func BenchmarkCount(b *testing.B) {
+	data := arrow.NewFloat(NormalData, &memory.Allocator{})
 	executetest.AggFuncBenchmarkHelper(
 		b,
 		new(transformations.CountAgg),
-		NormalData,
+		data,
 		int64(len(NormalData)),
 	)
 }