--- a/functions/transformations/percentile_test.go
+++ b/functions/transformations/percentile_test.go
@@ -5,9 +5,11 @@ import (
 	"testing"
 
 	"github.com/influxdata/flux"
+	"github.com/influxdata/flux/arrow"
 	"github.com/influxdata/flux/execute"
 	"github.com/influxdata/flux/execute/executetest"
 	"github.com/influxdata/flux/functions/transformations"
+	"github.com/influxdata/flux/memory"
 	"github.com/influxdata/flux/querytest"
 )
 
@@ -390,13 +392,14 @@ func TestPercentileSelector_Process(t *testing.T) {
 }
 
 func BenchmarkPercentile(b *testing.B) {
+	data := arrow.NewFloat(NormalData, &memory.Allocator{})
 	executetest.AggFuncBenchmarkHelper(
 		b,
 		&transformations.PercentileAgg{
 			Quantile:    0.9,
 			Compression: 1000,
 		},
-		NormalData,
+		data,
 		13.843815760607427,
 	)
 }