--- a/functions/transformations/sum_test.go
+++ b/functions/transformations/sum_test.go
@@ -4,8 +4,10 @@ import (
 	"testing"
 
 	"github.com/influxdata/flux"
+	"github.com/influxdata/flux/arrow"
 	"github.com/influxdata/flux/execute/executetest"
 	"github.com/influxdata/flux/functions/transformations"
+	"github.com/influxdata/flux/memory"
 	"github.com/influxdata/flux/querytest"
 )
 
@@ -28,10 +30,11 @@ func TestSum_Process(t *testing.T) {
 }
 
 func BenchmarkSum(b *testing.B) {
+	data := arrow.NewFloat(NormalData, &memory.Allocator{})
 	executetest.AggFuncBenchmarkHelper(
 		b,
 		new(transformations.SumAgg),
-		NormalData,
+		data,
 		10000816.96729983,
 	)
 }