… functions/tests/testdata/count.in.csv
… functions/tests/testdata/percentile_aggregate.in.csv
… functions/tests/testdata/percentile_tdigest.in.csv
… functions/tests/testdata/skew.in.csv
… functions/tests/testdata/spread.in.csv
… functions/tests/testdata/spread.out.csv
… functions/tests/testdata/stddev.in.csv
… functions/tests/testdata/sum.in.csv
… functions/tests/testdata/sum.out.csv