All aggregates will now use the arrow interface. They do not support nil
at the moment.

Fixes #491, #525, #511, #520, and #522.

This also implements part of #513. The other part uses the selector interface which is not included in this change.