@nathanielc 2018-12-14
@jsternberg Can you make sure that Flux tests exist for each one of the aggregates this fixes? As well as confirm that the spec is up to date? Since we have to touch every transformation the plan is to take the time to make sure they all have tests and are documentation.