@jsternberg 2018-12-19
I might be doing something wrong (or my hardware can't be consistent enough), but I'm not seeing a huge difference.

```
benchmark                                                      old ns/op      new ns/op      delta
BenchmarkCount-12                                              299            308            +3.01%
BenchmarkFirst-12                                              390            414            +6.15%
BenchmarkLast-12                                               24358515       24443051       +0.35%
BenchmarkMax-12                                                26627664       26522729       -0.39%
BenchmarkMin-12                                                25825030       25964603       +0.54%
BenchmarkSample-12                                             173622         168401         -3.01%
Benchmark_QueryEndToEnd/aggregate_window.flux-12               6096529        5705963        -6.41%
Benchmark_QueryEndToEnd/derivative.flux-12                     4861718        4440814        -8.66%
Benchmark_QueryEndToEnd/derivative_nonnegative.flux-12         4960126        4751617        -4.20%
Benchmark_QueryEndToEnd/difference.flux-12                     4557824        4480866        -1.69%
Benchmark_QueryEndToEnd/difference_one_value.flux-12           4444671        4376162        -1.54%
Benchmark_QueryEndToEnd/difference_panic.flux-12               8102767        8236249        +1.65%
Benchmark_QueryEndToEnd/distinct.flux-12                       5242532        5058229        -3.52%
Benchmark_QueryEndToEnd/drop_after_rename.flux-12              5905355        5819781        -1.45%
Benchmark_QueryEndToEnd/drop_newname_after.flux-12             5841888        5496010        -5.92%
Benchmark_QueryEndToEnd/drop_newname_before.flux-12            5781593        5830370        +0.84%
Benchmark_QueryEndToEnd/drop_unused.flux-12                    8815937        8535313        -3.18%
Benchmark_QueryEndToEnd/duplicate.flux-12                      4788151        5034555        +5.15%
Benchmark_QueryEndToEnd/filter_by_regex.flux-12                12646172       13242785       +4.72%
Benchmark_QueryEndToEnd/filter_by_regex_function.flux-12       9284144        9106073        -1.92%
Benchmark_QueryEndToEnd/filter_by_tags.flux-12                 12791397       12983752       +1.50%
Benchmark_QueryEndToEnd/group.flux-12                          15335798       14832382       -3.28%
Benchmark_QueryEndToEnd/group_by_field.flux-12                 5172118        5081580        -1.75%
Benchmark_QueryEndToEnd/group_by_irregular.flux-12             11186200       11920449       +6.56%
Benchmark_QueryEndToEnd/group_except.flux-12                   5493249        5529193        +0.65%
Benchmark_QueryEndToEnd/group_ungroup.flux-12                  10025128       9647920        -3.76%
Benchmark_QueryEndToEnd/highestAverage.flux-12                 4686340        4681161        -0.11%
Benchmark_QueryEndToEnd/highestCurrent.flux-12                 5139468        5510307        +7.22%
Benchmark_QueryEndToEnd/highestMax.flux-12                     4562761        4832739        +5.92%
Benchmark_QueryEndToEnd/histogram.flux-12                      4641052        4440404        -4.32%
Benchmark_QueryEndToEnd/histogram_quantile.flux-12             4855003        4746516        -2.23%
Benchmark_QueryEndToEnd/increase.flux-12                       4866515        4672236        -3.99%
Benchmark_QueryEndToEnd/influxFieldsAsCols.flux-12             5438738        5483550        +0.82%
Benchmark_QueryEndToEnd/join.flux-12                           20900610       21683995       +3.75%
Benchmark_QueryEndToEnd/join_across_measurements.flux-12       4678755083     4561001247     -2.52%
Benchmark_QueryEndToEnd/keep.flux-12                           5475521        4831230        -11.77%
Benchmark_QueryEndToEnd/keep_fn.flux-12                        11356151       12416107       +9.33%
Benchmark_QueryEndToEnd/keys.flux-12                           10835230       11247548       +3.81%
Benchmark_QueryEndToEnd/last.flux-12                           4256475        5318142        +24.94%
Benchmark_QueryEndToEnd/lowestAverage.flux-12                  5310368        6082435        +14.54%
Benchmark_QueryEndToEnd/lowestCurrent.flux-12                  5590192        5618330        +0.50%
Benchmark_QueryEndToEnd/lowestMin.flux-12                      5717622        5453022        -4.63%
Benchmark_QueryEndToEnd/mean.flux-12                           10682260       4977054        -53.41%
Benchmark_QueryEndToEnd/meta_query_fields.flux-12              9787813        10283772       +5.07%
Benchmark_QueryEndToEnd/meta_query_measurements.flux-12        5800209        5862437        +1.07%
Benchmark_QueryEndToEnd/multiple_range.flux-12                 4711351        4930705        +4.66%
Benchmark_QueryEndToEnd/parse_regex.flux-12                    8776606        8537676        -2.72%
Benchmark_QueryEndToEnd/percentile.flux-12                     10233206       10538959       +2.99%
Benchmark_QueryEndToEnd/pivot.flux-12                          6039618        5928549        -1.84%
Benchmark_QueryEndToEnd/pivot_fields.flux-12                   5813087        5769527        -0.75%
Benchmark_QueryEndToEnd/pivot_mean.flux-12                     7655504        7207902        -5.85%
Benchmark_QueryEndToEnd/pivot_task_test.flux-12                10208917       10387107       +1.75%
Benchmark_QueryEndToEnd/range.flux-12                          4353781        4225309        -2.95%
Benchmark_QueryEndToEnd/rename.flux-12                         5920960        5912140        -0.15%
Benchmark_QueryEndToEnd/rename_fn.flux-12                      11323471       11302393       -0.19%
Benchmark_QueryEndToEnd/rename_multiple.flux-12                5927302        5823263        -1.76%
Benchmark_QueryEndToEnd/select_measurement.flux-12             13933860       13878730       -0.40%
Benchmark_QueryEndToEnd/select_measurement_field.flux-12       14400439       13934988       -3.23%
Benchmark_QueryEndToEnd/selector_preserve_time.flux-12         5593016        5568518        -0.44%
Benchmark_QueryEndToEnd/show_all_tag_keys.flux-12              34354708       32751749       -4.67%
Benchmark_QueryEndToEnd/simple_max.flux-12                     9562341        9355821        -2.16%
Benchmark_QueryEndToEnd/sort.flux-12                           5350750        5241809        -2.04%
Benchmark_QueryEndToEnd/state_count.flux-12                    7506127        7897833        +5.22%
Benchmark_QueryEndToEnd/state_duration.flux-12                 8098396        7906718        -2.37%
Benchmark_QueryEndToEnd/string_sort.flux-12                    4865195        4771124        -1.93%
Benchmark_QueryEndToEnd/task_per_line.flux-12                  25725006       26091304       +1.42%
Benchmark_QueryEndToEnd/top.flux-12                            4889126        4853439        -0.73%
Benchmark_QueryEndToEnd/union.flux-12                          21650052       20800627       -3.92%
Benchmark_QueryEndToEnd/union_heterogeneous.flux-12            27811623       28348906       +1.93%
Benchmark_QueryEndToEnd/window.flux-12                         13902147       13741009       -1.16%
Benchmark_QueryEndToEnd/window_default_start_align.flux-12     5432266        5235253        -3.63%
Benchmark_QueryEndToEnd/window_generate_empty.flux-12          5943723        6023432        +1.34%
Benchmark_QueryEndToEnd/window_group_mean_ungroup.flux-12      6946770        6999504        +0.76%
Benchmark_QueryEndToEnd/window_offset.flux-12                  13890000       13814743       -0.54%
Benchmark_QueryEndToEnd/window_start_bound.flux-12             5288037        5124869        -3.09%

benchmark                                                      old allocs     new allocs     delta
Benchmark_QueryEndToEnd/aggregate_window.flux-12               42437          42391          -0.11%
Benchmark_QueryEndToEnd/derivative.flux-12                     30496          30593          +0.32%
Benchmark_QueryEndToEnd/derivative_nonnegative.flux-12         31462          31548          +0.27%
Benchmark_QueryEndToEnd/difference.flux-12                     29372          29363          -0.03%
Benchmark_QueryEndToEnd/difference_one_value.flux-12           29710          29747          +0.12%
Benchmark_QueryEndToEnd/difference_panic.flux-12               53208          53219          +0.02%
Benchmark_QueryEndToEnd/distinct.flux-12                       34710          34804          +0.27%
Benchmark_QueryEndToEnd/drop_after_rename.flux-12              41430          41466          +0.09%
Benchmark_QueryEndToEnd/drop_newname_after.flux-12             41062          41174          +0.27%
Benchmark_QueryEndToEnd/drop_newname_before.flux-12            41252          41580          +0.80%
Benchmark_QueryEndToEnd/drop_unused.flux-12                    58378          58632          +0.44%
Benchmark_QueryEndToEnd/duplicate.flux-12                      32445          32560          +0.35%
Benchmark_QueryEndToEnd/filter_by_regex.flux-12                97161          97580          +0.43%
Benchmark_QueryEndToEnd/filter_by_regex_function.flux-12       64760          64656          -0.16%
Benchmark_QueryEndToEnd/filter_by_tags.flux-12                 95348          95147          -0.21%
Benchmark_QueryEndToEnd/group.flux-12                          111196         111035         -0.14%
Benchmark_QueryEndToEnd/group_by_field.flux-12                 34555          34692          +0.40%
Benchmark_QueryEndToEnd/group_by_irregular.flux-12             88243          88361          +0.13%
Benchmark_QueryEndToEnd/group_except.flux-12                   38764          38855          +0.23%
Benchmark_QueryEndToEnd/group_ungroup.flux-12                  70869          71044          +0.25%
Benchmark_QueryEndToEnd/highestAverage.flux-12                 29040          29126          +0.30%
Benchmark_QueryEndToEnd/highestCurrent.flux-12                 37281          37415          +0.36%
Benchmark_QueryEndToEnd/highestMax.flux-12                     30070          30234          +0.55%
Benchmark_QueryEndToEnd/histogram.flux-12                      27917          27972          +0.20%
Benchmark_QueryEndToEnd/histogram_quantile.flux-12             31873          31823          -0.16%
Benchmark_QueryEndToEnd/increase.flux-12                       30660          30823          +0.53%
Benchmark_QueryEndToEnd/influxFieldsAsCols.flux-12             36717          36696          -0.06%
Benchmark_QueryEndToEnd/join.flux-12                           176666         176921         +0.14%
Benchmark_QueryEndToEnd/join_across_measurements.flux-12       864241         864843         +0.07%
Benchmark_QueryEndToEnd/keep.flux-12                           33928          33996          +0.20%
Benchmark_QueryEndToEnd/keep_fn.flux-12                        77471          77667          +0.25%
Benchmark_QueryEndToEnd/keys.flux-12                           79583          79350          -0.29%
Benchmark_QueryEndToEnd/last.flux-12                           26617          26695          +0.29%
Benchmark_QueryEndToEnd/lowestAverage.flux-12                  36285          36305          +0.06%
Benchmark_QueryEndToEnd/lowestCurrent.flux-12                  37291          37426          +0.36%
Benchmark_QueryEndToEnd/lowestMin.flux-12                      37308          37478          +0.46%
Benchmark_QueryEndToEnd/mean.flux-12                           77570          32148          -58.56%
Benchmark_QueryEndToEnd/meta_query_fields.flux-12              70371          70913          +0.77%
Benchmark_QueryEndToEnd/meta_query_measurements.flux-12        41574          41446          -0.31%
Benchmark_QueryEndToEnd/multiple_range.flux-12                 33014          33136          +0.37%
Benchmark_QueryEndToEnd/parse_regex.flux-12                    56487          56630          +0.25%
Benchmark_QueryEndToEnd/percentile.flux-12                     75958          75846          -0.15%
Benchmark_QueryEndToEnd/pivot.flux-12                          41976          42173          +0.47%
Benchmark_QueryEndToEnd/pivot_fields.flux-12                   41724          41892          +0.40%
Benchmark_QueryEndToEnd/pivot_mean.flux-12                     56545          56562          +0.03%
Benchmark_QueryEndToEnd/pivot_task_test.flux-12                69771          69499          -0.39%
Benchmark_QueryEndToEnd/range.flux-12                          27109          26943          -0.61%
Benchmark_QueryEndToEnd/rename.flux-12                         41564          41679          +0.28%
Benchmark_QueryEndToEnd/rename_fn.flux-12                      76296          76303          +0.01%
Benchmark_QueryEndToEnd/rename_multiple.flux-12                41279          41377          +0.24%
Benchmark_QueryEndToEnd/select_measurement.flux-12             97216          97011          -0.21%
Benchmark_QueryEndToEnd/select_measurement_field.flux-12       103456         103713         +0.25%
Benchmark_QueryEndToEnd/selector_preserve_time.flux-12         38997          39047          +0.13%
Benchmark_QueryEndToEnd/show_all_tag_keys.flux-12              159457         159658         +0.13%
Benchmark_QueryEndToEnd/simple_max.flux-12                     65533          65759          +0.34%
Benchmark_QueryEndToEnd/sort.flux-12                           34626          34754          +0.37%
Benchmark_QueryEndToEnd/state_count.flux-12                    50040          50087          +0.09%
Benchmark_QueryEndToEnd/state_duration.flux-12                 49854          50201          +0.70%
Benchmark_QueryEndToEnd/string_sort.flux-12                    29175          29428          +0.87%
Benchmark_QueryEndToEnd/task_per_line.flux-12                  226317         225757         -0.25%
Benchmark_QueryEndToEnd/top.flux-12                            30822          30744          -0.25%
Benchmark_QueryEndToEnd/union.flux-12                          174397         174285         -0.06%
Benchmark_QueryEndToEnd/union_heterogeneous.flux-12            236421         236374         -0.02%
Benchmark_QueryEndToEnd/window.flux-12                         106464         106344         -0.11%
Benchmark_QueryEndToEnd/window_default_start_align.flux-12     34310          34563          +0.74%
Benchmark_QueryEndToEnd/window_generate_empty.flux-12          41574          41633          +0.14%
Benchmark_QueryEndToEnd/window_group_mean_ungroup.flux-12      50803          50807          +0.01%
Benchmark_QueryEndToEnd/window_offset.flux-12                  109685         109920         +0.21%
Benchmark_QueryEndToEnd/window_start_bound.flux-12             34438          34535          +0.28%

benchmark                                                      old bytes     new bytes     delta
Benchmark_QueryEndToEnd/aggregate_window.flux-12               2971113       2967949       -0.11%
Benchmark_QueryEndToEnd/derivative.flux-12                     2104936       2107853       +0.14%
Benchmark_QueryEndToEnd/derivative_nonnegative.flux-12         2154519       2162736       +0.38%
Benchmark_QueryEndToEnd/difference.flux-12                     2035806       2031878       -0.19%
Benchmark_QueryEndToEnd/difference_one_value.flux-12           2047924       2050146       +0.11%
Benchmark_QueryEndToEnd/difference_panic.flux-12               3577033       3579712       +0.07%
Benchmark_QueryEndToEnd/distinct.flux-12                       2450168       2459468       +0.38%
Benchmark_QueryEndToEnd/drop_after_rename.flux-12              2880156       2885480       +0.18%
Benchmark_QueryEndToEnd/drop_newname_after.flux-12             2850144       2862000       +0.42%
Benchmark_QueryEndToEnd/drop_newname_before.flux-12            2869089       2888337       +0.67%
Benchmark_QueryEndToEnd/drop_unused.flux-12                    3910930       3924314       +0.34%
Benchmark_QueryEndToEnd/duplicate.flux-12                      2240600       2251855       +0.50%
Benchmark_QueryEndToEnd/filter_by_regex.flux-12                6521051       6548586       +0.42%
Benchmark_QueryEndToEnd/filter_by_regex_function.flux-12       4342475       4328933       -0.31%
Benchmark_QueryEndToEnd/filter_by_tags.flux-12                 6393253       6368249       -0.39%
Benchmark_QueryEndToEnd/group.flux-12                          7425362       7414814       -0.14%
Benchmark_QueryEndToEnd/group_by_field.flux-12                 2366483       2376890       +0.44%
Benchmark_QueryEndToEnd/group_by_irregular.flux-12             5817445       5833257       +0.27%
Benchmark_QueryEndToEnd/group_except.flux-12                   2649146       2659409       +0.39%
Benchmark_QueryEndToEnd/group_ungroup.flux-12                  4796018       4812859       +0.35%
Benchmark_QueryEndToEnd/highestAverage.flux-12                 1885698       1893877       +0.43%
Benchmark_QueryEndToEnd/highestCurrent.flux-12                 2470053       2479339       +0.38%
Benchmark_QueryEndToEnd/highestMax.flux-12                     1958422       1969266       +0.55%
Benchmark_QueryEndToEnd/histogram.flux-12                      1880019       1880899       +0.05%
Benchmark_QueryEndToEnd/histogram_quantile.flux-12             2174970       2170328       -0.21%
Benchmark_QueryEndToEnd/increase.flux-12                       2138716       2150279       +0.54%
Benchmark_QueryEndToEnd/influxFieldsAsCols.flux-12             2592367       2591858       -0.02%
Benchmark_QueryEndToEnd/join.flux-12                           11470493      11480089      +0.08%
Benchmark_QueryEndToEnd/join_across_measurements.flux-12       81749800      81830408      +0.10%
Benchmark_QueryEndToEnd/keep.flux-12                           2317761       2323151       +0.23%
Benchmark_QueryEndToEnd/keep_fn.flux-12                        5062618       5066708       +0.08%
Benchmark_QueryEndToEnd/keys.flux-12                           5456271       5441538       -0.27%
Benchmark_QueryEndToEnd/last.flux-12                           1853163       1852154       -0.05%
Benchmark_QueryEndToEnd/lowestAverage.flux-12                  2400590       2402220       +0.07%
Benchmark_QueryEndToEnd/lowestCurrent.flux-12                  2473799       2477349       +0.14%
Benchmark_QueryEndToEnd/lowestMin.flux-12                      2473435       2481768       +0.34%
Benchmark_QueryEndToEnd/mean.flux-12                           5345294       2203433       -58.78%
Benchmark_QueryEndToEnd/meta_query_fields.flux-12              4765549       4807680       +0.88%
Benchmark_QueryEndToEnd/meta_query_measurements.flux-12        2897394       2884868       -0.43%
Benchmark_QueryEndToEnd/multiple_range.flux-12                 2240106       2249808       +0.43%
Benchmark_QueryEndToEnd/parse_regex.flux-12                    3841945       3852210       +0.27%
Benchmark_QueryEndToEnd/percentile.flux-12                     5102743       5097631       -0.10%
Benchmark_QueryEndToEnd/pivot.flux-12                          2836741       2855361       +0.66%
Benchmark_QueryEndToEnd/pivot_fields.flux-12                   2835940       2846653       +0.38%
Benchmark_QueryEndToEnd/pivot_mean.flux-12                     3838944       3833639       -0.14%
Benchmark_QueryEndToEnd/pivot_task_test.flux-12                4525395       4502644       -0.50%
Benchmark_QueryEndToEnd/range.flux-12                          1865786       1850042       -0.84%
Benchmark_QueryEndToEnd/rename.flux-12                         2876682       2881495       +0.17%
Benchmark_QueryEndToEnd/rename_fn.flux-12                      5033870       5038620       +0.09%
Benchmark_QueryEndToEnd/rename_multiple.flux-12                2894927       2894869       -0.00%
…